core: split the XML engine into edition shims
Split each XML engine file into a thin, edition-neutral shim plus a
single-track _core implementation, so an alternate engine variant can be
loaded through the existing EDITION flag without editing the shared files
inline.

editor_base.rb, export.rb, fcpx.rb and fcp7.rb become shims that load
<name>_core.rb — or <name>_pro.rb, when that file is present — through the
new ButterCut.engine_variant helper. export.rb keeps the CLI entrypoint.
Add ButterCut.pro? and ButterCut.core? predicates.

Behavior-preserving: core always resolves to the existing single-track
_core code. Full suite green (311 examples), plus a spec pinning the
edition-resolution contract.
